+++ neo4j/_work/__init__.py
@@ -15,20 +15,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# TODO: 6.0 - remove this file
 
-
-from .._meta import deprecation_warn
-from .._work import ResultSummary
+from .query import (
+    Query,
+    unit_of_work,
+)
+from .summary import (
+    ResultSummary,
+    SummaryCounters,
+)
 
 
 __all__ = [
+    "Query",
     "ResultSummary",
+    "SummaryCounters",
+    "unit_of_work",
 ]
-
-deprecation_warn(
-    "The module `neo4j.work.summary` was made internal and will "
-    "no longer be available for import in future versions. "
-    "Everything from there should be imported directly from 'neo4j'.",
-    stack_level=2
-)